fix: remove circular dependency by moving parse_s3 method to its own util file (#5430)
* fix: remove circular dependency by moving parse_s3 method to its own util file

* add missing unit tests file
